pyproject.toml
[project]
name = "mypub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bibtexparser>=1.4.4",
    "google-genai>=1.62.0",
    "openai>=2.17.0",
    "requests>=2.32.5",
    "toml>=0.10.2",
]
